Fix the issue with expanding/collapsing balance cell
The problem was in initializing BalanceViewModel. When user changes
theme the BalanceFragment is recreated. It triggered init method of
BalanceViewModel for the second time. Then the same BalanceViewModel
initialized with a new set of Presenter and Interactor. As a result,
there were 2 independent balance presenters that trigger one single
BalanceViewModel with different expanded item. That is why there were
flicker.