Fix the issue with delay in applying font to toolbar title
I've found out the reason is that toolbar uses the downloadable font.
In some cases the title is shown using default font and then applies
the specified one. But at the same time, if the TextView uses the same
downloadable font it applies it without any delay.

Including font in app resolves the issue.