Zcash: Remove GlobalScope in favor of synchronizer's scope.
Using GlobalScope usually results in memory leaks and code running when the app is inactive. The Synchronizer can be tied to any scope but it also has it's own lifecycle that ends with 'stop' is called. So that scope can be used in the worst case when there is no over-arching scope to use instead. Typically the activity lifecycle or a subcomponent scope from Dagger would be used, instead. Also, transactions should be accessed after starting the synchronizer to allow things like database migrations to run, first. Although the SDK was recently changed so that this isn't strictly required, it is probably safer for subscription to happen after start, rather than before. Also subscribing during 'init' can make loading the app a little less responsive.