Zcash: Show sync percentage.
Right now sync percentage does not show unless lastBlockDate exists. Those two values could probably be decoupled so that percentage can be shown even if there is no date. For now, we can use a hardcoded date on for Zcash, since syncing can take a long time and having a percentage is helpful.