Add option to pause sync of tx and fee data
Signing with a hardware wallet could take some time.
We need to pause sync of transaction estimaded fees
and other data in case it updated while signing with a hardware wallet.

After the transaction has been sent to the hardware wallet for singing
we will pause background sync of transaction parameters